config/routes.rb: remove `.freeze` from EZBORROW_URL assignment rvalue [+]
`freeze` called on the resulting string value does not actually turn the EZBORROW_URL into a constant.
The variable can still be changed just as before.  Freezing the string does render it (meaning the rvalue,
not the variable it's being assigned to) immutable which can allow for certain small performance gains,
but they would only be realized under certain conditions, which would not appear to be present here.
Removing freeze to prevent any future confusion, as some devs might think that EZBORROW_URL is
protected by it.